+++ src/main/java/io/github/jamalam360/tb/bone/drops/TotallyBalancedBoneDropsInit.java
@@ -24,29 +24,25 @@
 
 package io.github.jamalam360.tb.bone.drops;
 
-import com.google.gson.Gson;
-import com.google.gson.JsonElement;
 import io.github.jamalam360.jamlib.config.JamLibConfig;
 import net.fabricmc.api.ModInitializer;
-import net.fabricmc.fabric.api.resource.IdentifiableResourceReloadListener;
-import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
+import net.minecraft.enchantment.EnchantmentHelper;
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.EntityType;
+import net.minecraft.entity.LivingEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
-import net.minecraft.resource.JsonDataLoader;
-import net.minecraft.resource.ResourceManager;
-import net.minecraft.resource.ResourceType;
+import net.minecraft.tag.TagKey;
 import net.minecraft.util.Identifier;
-import net.minecraft.util.profiler.Profiler;
+import net.minecraft.util.registry.Registry;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
 import java.util.Random;
 
 public class TotallyBalancedBoneDropsInit implements ModInitializer {
     public static final String MOD_NAME = "Totally Balanced Bone Drops";
+    public static final TagKey<EntityType<?>> BLACKLIST = TagKey.of(Registry.ENTITY_TYPE_KEY, new Identifier("tb-bone-drops", "blacklist"));
     private static final Logger LOGGER = getLogger("Initializer");
     private static final Random RANDOM = new Random();
 
@@ -54,7 +50,6 @@ public class TotallyBalancedBoneDropsInit implements ModInitializer {
     public void onInitialize() {
         LOGGER.info("Initializing " + MOD_NAME + "...");
         JamLibConfig.init("tb-bone-drops", Config.class);
-        ResourceManagerHelper.get(ResourceType.SERVER_DATA).registerReloadListener(new DataLoader());
     }
 
     public static Logger getLogger(String name) {
@@ -65,8 +60,8 @@ public class TotallyBalancedBoneDropsInit implements ModInitializer {
         VERY_RARE(0.05D, 1),
         RARE(0.1D, 2),
         COMMON(0.2D, 2),
-        VERY_COMMON(0.4D, 2),
-        ABUNDANT(0.7D, 3);
+        VERY_COMMON(0.35D, 2),
+        ABUNDANT(0.55D, 3);
 
         private final double chance;
         private final int rolls;
@@ -76,40 +71,22 @@ public class TotallyBalancedBoneDropsInit implements ModInitializer {
             this.rolls = rolls;
         }
 
-        public ItemStack getDrop() {
+        public ItemStack getDrop(Entity attacker) {
             int count = 0;
 
-            for (int i = 0; i < this.rolls; i++) {
-                if (RANDOM.nextDouble() < this.chance) {
+            double mult = 1;
+
+            if (attacker instanceof LivingEntity living) {
+                mult = 1 + EnchantmentHelper.getLooting(living) * 0.3;
+            }
+
+            for (int i = 0; i < Math.ceil(this.rolls * mult); i++) {
+                if (RANDOM.nextDouble() < (this.chance * mult)) {
                     count++;
                 }
             }
 
             return new ItemStack(Items.BONE, count);
-        }
-    }
-
-    public static class DataLoader extends JsonDataLoader implements IdentifiableResourceReloadListener {
-        public static List<Identifier> BLACKLIST = new ArrayList<>();
-
-        public DataLoader() {
-            super(new Gson(), "tb_bone_drops");
-        }
-
-        @Override
-        public void apply(Map<Identifier, JsonElement> prepared, ResourceManager manager, Profiler profiler) {
-            BLACKLIST.clear();
-
-            prepared.forEach((id, element) -> {
-                for (var entry : element.getAsJsonObject().get("blacklist").getAsJsonArray()) {
-                    BLACKLIST.add(new Identifier(entry.getAsString()));
-                }
-            });
-        }
-
-        @Override
-        public Identifier getFabricId() {
-            return new Identifier("tb-bone-drops", "tb_bone_drops_data");
         }
     }
 }