+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ### Features
 
-- Support 1.18 by using a `java.util.Random` in place of a `RandomGenerator`.
+- Move the custom entity blacklist to a tag (`totally-balanced-bone-drops:blacklist`).
+- Increase the chance of dropping a bone if looting is present on the attackers' weapon.
+
+### Fixes
+
+- Decrease chances of dropping bones for all config options.
 
 Closed issues: None.
 
-[Full Changelog](https://github.com/JamCoreModding/TotallyBalancedBoneDrops/compare/1.1.0...1.1.1)
+[Full Changelog](https://github.com/JamCoreModding/TotallyBalancedBoneDrops/compare/1.1.1...1.2.0)